Move InterfaceFormField to separate component file
